standardize uint

+++ contracts/bondingcurve/BondingCurve.sol
@@ -11,13 +11,13 @@ import "../pcv/PCVSplitter.sol";
 /// @author Fei Protocol
 abstract contract BondingCurve is IBondingCurve, OracleRef, PCVSplitter {
     using Decimal for Decimal.D256;
-    using Roots for uint256;
+    using Roots for uint;
 
-	uint256 public override scale;
-	uint256 public override totalPurchased; // FEI_b for this curve
+	uint public override scale;
+	uint public override totalPurchased; // FEI_b for this curve
 
-	uint256 public override buffer = 100;
-	uint256 public constant BUFFER_GRANULARITY = 10_000;
+	uint public override buffer = 100;
+	uint public constant BUFFER_GRANULARITY = 10_000;
 
 	/// @notice constructor
 	/// @param _scale the Scale target where peg fixes
@@ -26,10 +26,10 @@ abstract contract BondingCurve is IBondingCurve, OracleRef, PCVSplitter {
 	/// @param _ratios the ratios for the PCVSplitter
 	/// @param _oracle the UniswapOracle to reference
 	constructor(
-		uint256 _scale, 
+		uint _scale, 
 		address _core, 
 		address[] memory _pcvDeposits, 
-		uint256[] memory _ratios, 
+		uint[] memory _ratios, 
 		address _oracle
 	) public
 		OracleRef(_core, _oracle)
@@ -38,18 +38,18 @@ abstract contract BondingCurve is IBondingCurve, OracleRef, PCVSplitter {
 		_setScale(_scale);
 	}
 
-	function setScale(uint256 _scale) public override onlyGovernor {
+	function setScale(uint _scale) public override onlyGovernor {
 		_setScale(_scale);
 		emit ScaleUpdate(_scale);
 	}
 
-	function setBuffer(uint256 _buffer) public override onlyGovernor {
+	function setBuffer(uint _buffer) public override onlyGovernor {
 		require(_buffer < BUFFER_GRANULARITY, "BondingCurve: Buffer exceeds or matches granularity");
 		buffer = _buffer;
 		emit BufferUpdate(_buffer);
 	}
 
-	function setAllocation(address[] memory allocations, uint256[] memory ratios) public override onlyGovernor {
+	function setAllocation(address[] memory allocations, uint[] memory ratios) public override onlyGovernor {
 		_setAllocation(allocations, ratios);
 	}
 
@@ -64,25 +64,25 @@ abstract contract BondingCurve is IBondingCurve, OracleRef, PCVSplitter {
 		return peg().div(_getBondingCurvePriceMultiplier());
 	}
 
-	function getAmountOut(uint256 amountIn) public view override returns (uint256 amountOut) {
-		uint256 adjustedAmount = getAdjustedAmount(amountIn);
+	function getAmountOut(uint amountIn) public view override returns (uint amountOut) {
+		uint adjustedAmount = getAdjustedAmount(amountIn);
 		if (atScale()) {
 			return _getBufferAdjustedAmount(adjustedAmount);
 		}
 		return _getBondingCurveAmountOut(adjustedAmount);
 	}
 
-	function getAveragePrice(uint256 amountIn) public view override returns (Decimal.D256 memory) {
-		uint256 adjustedAmount = getAdjustedAmount(amountIn);
-		uint256 amountOut = getAmountOut(amountIn);
+	function getAveragePrice(uint amountIn) public view override returns (Decimal.D256 memory) {
+		uint adjustedAmount = getAdjustedAmount(amountIn);
+		uint amountOut = getAmountOut(amountIn);
 		return Decimal.ratio(adjustedAmount, amountOut);
 	}
 
-	function getAdjustedAmount(uint256 amountIn) internal view returns (uint256) {
+	function getAdjustedAmount(uint amountIn) internal view returns (uint) {
 		return peg().mul(amountIn).asUint256();
 	}
 
-	function _purchase(uint256 amountIn, address to) internal returns (uint256 amountOut) {
+	function _purchase(uint amountIn, address to) internal returns (uint amountOut) {
 	 	updateOracle();
 		
 	 	amountOut = getAmountOut(amountIn);
@@ -96,24 +96,24 @@ abstract contract BondingCurve is IBondingCurve, OracleRef, PCVSplitter {
 		return amountOut;
 	}
 
-	function incrementTotalPurchased(uint256 amount) internal {
+	function incrementTotalPurchased(uint amount) internal {
 		totalPurchased += amount;
 	}
 
-	function _setScale(uint256 _scale) internal {
+	function _setScale(uint _scale) internal {
 		scale = _scale;
 	}
 
 	function _getBondingCurvePriceMultiplier() internal view virtual returns(Decimal.D256 memory);
 
-	function _getBondingCurveAmountOut(uint256 adjustedAmountIn) internal view virtual returns(uint256);
+	function _getBondingCurveAmountOut(uint adjustedAmountIn) internal view virtual returns(uint);
 
 	function _getBuffer() internal view returns(Decimal.D256 memory) {
 		uint granularity = BUFFER_GRANULARITY;
 		return Decimal.ratio(granularity - buffer, granularity);
 	} 
 
-	function _getBufferAdjustedAmount(uint256 amountIn) internal view returns(uint256) {
+	function _getBufferAdjustedAmount(uint amountIn) internal view returns(uint) {
 		return _getBuffer().mul(amountIn).asUint256();
 	}
 }
