+++ contracts/pcv/EthUniswapPCVController.sol
@@ -105,7 +105,7 @@ contract EthUniswapPCVController is UniRef, IUniswapPCVController {
 		weth.deposit{value: amount}();
 		weth.transfer(address(pair), amount);
 
-		(uint256 amount0Out, uint256 amount1Out) = pair.token0() == address(weth) ? (uint256(0), amountOut) : (amountOut, uint256(0));
+		(uint amount0Out, uint amount1Out) = pair.token0() == address(weth) ? (uint(0), amountOut) : (amountOut, uint(0));
 		pair.swap(amount0Out, amount1Out, address(this), new bytes(0));
 	}
