+++ contracts/pcv/IPCVDeposit.sol
@@ -8,15 +8,15 @@ interface IPCVDeposit {
 
     /// @notice deposit tokens into the PCV allocation
     /// @param amount of tokens deposited
-    function deposit(uint256 amount) external payable;
+    function deposit(uint amount) external payable;
 
     // PCV Controller only state updating api
 
     /// @notice withdraw tokens from the PCV allocation
     /// @param amount of tokens withdrawn
     /// @param to the address to send PCV to
-    function withdraw(address to, uint256 amount) external;
+    function withdraw(address to, uint amount) external;
 
     /// @notice returns total value of PCV in the Deposit
-    function totalValue() external view returns(uint256);
+    function totalValue() external view returns(uint);
 }
